refactor(api): refactor shared test folder
